ContextHub: Avoid reference leaks
The ContextHub HAL was leaking references in passOnOsResponse. An
example of this is when a sensor hub reset occurs. If many of these
resets happen, the local reference pool is exhausted and a crash
in the HAL occurs.

Bug: 30706325
Change-Id: Ica5328137a7033f512d7fb1ccf6fa138eab55091
(cherry picked from commit b9a1d96454efb40ac6aa36028b00cc9d6839f2c8)