Clear WS.mDestroying on AWT.clearAnimatingFlags
We call AppWindowToken.clearAnimatingFlags() when an app token is
no longer considered exiting since it is now part of the opening
app transition. To the same effect we should no longer consider
its windows destroying so they can be factored in when decided
what to do during the transition like picking the right wallpaper
target.

Bug: 30790402
Change-Id: Ide94af72a38bd4a955abb122b47f921fd3f70652