Update JavaDoc for IMS#enableHardwareAcceleration
This makes it clear that on Android L+ devices IME developers basically
no longer need to call InputMethodService#enableHardwareAcceleration(),
because it is deprecated in Android L[1][2].  On L+ devices what IME
developers need to do is to make sure that their IMEs are compatible
with both hardware and software rendering so that the system can choose
the best rendering method.

This is basically a code clean-up that should not change any actual
behavior.

 [1]: I906d48cb07af50fa02dae4f4ecdb5e65211fc6ef
      f0803cc409d45fc4246e0dbd95c6b9c86839e58e
 [2]: I126ea823eb2eeabb29d30999dec9c9411b33b315
      2afed4e51a506d63808011bc8c16fa5f89e84d35

Bug: 31202732
Change-Id: I416669b172737e3eb3d77165dcea50289d207514