Implement untetherAll as calls to stopTethering
(cherry-pick of 62dd7ce9db3e965b22317c148d651caf44d84b66)

Calling untether() on an interface name leaves the system
in an inconsistent state, since this is an API for use by
network subsystems to start or stop the tethering of a prepared
downstream network interface.  The proper way to disable tethering
is to call stopTethering() with an appropriate technology.

untetherAll() is called in parts of the system UI to disable
all tethering on system state changes (e.g. on the enabling
of data saver).

Bug: 31405407
Test: Enabling data saver now correctly disables tethering
Change-Id: I8e5724a31e015721910a8d278e100daf697c34d7