Remove WifiManager#get/setFrequencyBand
These methods are marked @hide and only partially implemented by
the WifiService.  In general, it seems wrong to strictly prefer
one WiFi band over another for all WiFi operations.  To reduce
implementation complexity, cause setFrequencyBand to become a noop
and getFrequencyBand to reflect the underlying reality, which is
that we will always operate on all supported bands.

Bug: 31148223
Test: Compiles
Change-Id: Iccff517915f7d5ed7e96bf56e4f195b2cda80f0a