Prevent any rotation while seamless rotation is pending.
Various errors occur when using even the normal rotation animation while
seamless rotation is pending. So we just defer the rotation like we do
for the normal animation. Since we are doing this, we need to track
when seamless rotation finishes so we can perform a post-rotate
rotation if required.

Bug: 31749456
Change-Id: I99f189306c690ce868496460e9ca7dcc95e4ccdc