Dreams: Fix lockscreen flicker
Fixes a bug where dreams (and other SHOW_WHEN_LOCKED activites) would
show a flicker of the lockscreen when the device locked while they are
showing.

This was caused because the lockscreen drives locking the device by 
making itself expand, but does not immediately know that it should hide
itself because the occluded state is only pushed when the lockscreen is
showing.

To fix this, we simply push the occluded state also when the device
isn't locked - that way, the status bar can go into the appropriate
state immediately upon locking.

Change-Id: Ie03bfcfe902fc562d76a773d3fc3371359f63563
Fixes: 29555358
Test: adb shell am start -n com.android.systemui/.Somnambulator