Not blocking focus on action bar on tab
When the focus jumps from action bar to
outside of it, action bar disables focusing
on it so that tabbing couldn’t get back to it.
Currently, it does the check that the focus
has left the bar from onFocusChange(), when
the internal data structures are in process
of updating the current focus; as a result, the
check may report a wrong result, and the
focus will be blocked while it’s actually is still in
the action bar. This caused the action bar to
lose focus.
The fix makes the check in a posted action,
fixing the issue.

Bug: 32087180
Test: Manual test that the issue doesn’t repro.
Change-Id: I67069d2ae58f2843e6fcc8c705943324cc1533f4