Power manager always needs to allow wake locks for services, receivers
The change to not allow wake locks for idle apps was a little too
aggressive, since we allow apps that are currently executing services
or receivers to be idle.  (Idle means that it hasn't been in the
foreground for X amount of time, but services and receivers don't count
as being in the foreground.)

So this takes in to account the actual procstate of the uid to determine
if it should actually respect the current idle state.

Test: manual

Change-Id: I32949124b37875b13233478719abccb8369e47f6